Languages/Python/SSGs/slideshow/requirements.txt
python-pptx
jinja2